+++ pservice/bin/ps-start.sh
@@ -20,23 +20,22 @@ if [[ $PY3_VERSION -lt 5 ]]; then
     exit
 fi
 
-F_USAGE='-c|--count services --cfile config -l|--loglevel [debug|info|warn]'
-F_LOGLEVEL='info'
-F_COUNT=1
-F_OUTPUTDIR=''
-F_BASENAME='pservice'
-
 F_SERVICEHOME="$( cd -P "$( dirname ${BASH_SOURCE[0]} )/.." && pwd )"
-F_LOGDIR=$F_SERVICEHOME/logs
-F_CONFDIR=$F_SERVICEHOME/etc
+source ${F_SERVICEHOME}/bin/common.sh # from ../../eservice/bin/
+
+F_USAGE='-c|--count services -b|--base name --ledger url -o|--output dir --clean -l|--loglevel [debug|info|warn]'
+F_COUNT=1
+F_BASENAME='pservice'
 F_LEDGERURL=''
+F_OUTPUTDIR=''
 F_CLEAN='no'
+F_LOGLEVEL='info'
 
 # -----------------------------------------------------------------
 # Process command line arguments
 # -----------------------------------------------------------------
 TEMP=`getopt -o b:c:l:o: --long base:,clean,count:,help,loglevel:,output:,ledger: \
-     -n 'launch.sh' -- "$@"`
+     -n 'ps-start.sh' -- "$@"`
 
 if [ $? != 0 ] ; then echo "Terminating..." >&2 ; exit 1 ; fi
 
@@ -46,15 +45,23 @@ while true ; do
         -b|--base) F_BASENAME="$2" ; shift 2 ;;
         --clean) F_CLEAN="yes" ; shift 1 ;;
         -c|--count) F_COUNT="$2" ; shift 2 ;;
-        -l|--loglevel) F_LOGLEVEL="$2" ; shift 2 ;;
         --ledger) F_LEDGERURL="--ledger $2" ; shift 2 ;;
+        -l|--loglevel) F_LOGLEVEL="$2" ; shift 2 ;;
         -o|--output) F_OUTPUTDIR="$2" ; shift 2 ;;
         --help) echo $F_USAGE ; exit 1 ;;
-    --) shift ; break ;;
-    *) echo "Internal error!" ; exit 1 ;;
+	--) shift ; break ;;
+	*) echo "Internal error!" ; exit 1 ;;
     esac
 done
 
+# (1) do not start if service already running
+pgrepf  "\bpservice .* --config ${F_BASENAME}[0-9].toml\b"
+if [ $? == 0 ] ; then
+    echo existing provisioning services detected, please shutdown first
+    exit 1
+fi
+
+# (2) prepare output
 if [ "$F_OUTPUTDIR" != "" ] ; then
     mkdir -p $F_OUTPUTDIR
     rm -f $F_OUTPUTDIR/*
@@ -63,6 +70,7 @@ else
     OFILE=/dev/null
 fi
 
+# (3) start services asynchronously
 for index in `seq 1 $F_COUNT` ; do
     IDENTITY="${F_BASENAME}$index"
     echo start provisioning service $IDENTITY
@@ -84,4 +92,17 @@ for index in `seq 1 $F_COUNT` ; do
     pservice --identity ${IDENTITY} --config ${IDENTITY}.toml enclave.toml --config-dir ${F_CONFDIR} ${F_LEDGERURL} \
              --loglevel ${F_LOGLEVEL} --logfile ${F_LOGDIR}/${IDENTITY}.log 2> $EFILE > $OFILE &
 
+done
+
+# (4) wait for successfull start of the services
+for index in `seq 1 $F_COUNT` ; do
+    IDENTITY="${F_BASENAME}$index"
+    echo waiting for startup completion of provisioning service $IDENTITY
+
+    url="$(get_url_base $IDENTITY)/info" || { echo "no url found for provisioning service"; exit 1; }
+    resp=$(${CURL_CMD} ${url})
+    if [ $? != 0 ] || [ $resp != "200" ]; then
+	echo "provisioning service $IDENTITY not properly running"
+	exit 1
+    fi
 done
